reformat rundowns into one
